# internal/server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tofi API
  version: 0.1.7
  description: |
    Tofi is an AI App Engine. Create AI-powered apps that run on a schedule
    or on-demand, with results delivered to your preferred channels.

    ## Authentication
    All endpoints (except login) require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Quick Start

    **Step 1: Login**
    ```bash
    curl -X POST https://tofi.sentiosurge.com/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"username": "your-username", "password": "your-password"}'
    ```
    Save the `token` from the response.

    **Step 2: Check your status**
    ```bash
    curl https://tofi.sentiosurge.com/api/v1/auth/me \
      -H "Authorization: Bearer YOUR_TOKEN"
    ```
    This shows your configured AI providers. If `ai_keys` is empty, go to Step 3.

    **Step 3: Set your AI key**
    ```bash
    curl -X PUT https://tofi.sentiosurge.com/api/v1/user/settings/ai-key \
      -H "Authorization: Bearer YOUR_TOKEN" \
      -H "Content-Type: application/json" \
      -d '{"provider": "openai", "api_key": "sk-..."}'
    ```
    Supported providers: `openai`, `anthropic`, `gemini`, `deepseek`, `groq`, `openrouter`

    **Step 4: Start chatting**
    ```bash
    # Create a session
    curl -X POST https://tofi.sentiosurge.com/api/v1/chat/sessions \
      -H "Authorization: Bearer YOUR_TOKEN" \
      -H "Content-Type: application/json" \
...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      summary: Login
      description: |
        Authenticate with username and password to receive a JWT token.
        Use this token in the Authorization header for all other endpoints.
        No authentication required for this endpoint.
      operationId: login
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
            example:
              username: "alice"
              password: "your-password"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token (valid for 365 days)
                  username:
                    type: string
                  role:
                    type: string
                    enum: [admin, user]
              example:
                token: "eyJhbGciOiJIUzI1NiIs..."
                username: "alice"
                role: "user"
        '401':
          description: Invalid username or password
...
  /api/v1/chat/sessions:
    post:
      summary: Create Chat Session
      description: |
        Create a new chat session. All fields are optional.
        Use the returned `id` to send messages.
      operationId: createChatSession
      tags: [Chat]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                  description: LLM model to use (auto-detected if omitted)
                scope:
                  type: string
                  description: Session scope (empty for general chat)
                skills:
                  type: array
                  items:
                    type: string
                  description: Skill names to enable
            example:
              model: "claude-sonnet-4-20250514"
              skills: ["web-search"]
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  scope:
                    type: string
                  model:
                    type: string
                  skills:
                    type: array
                    items:
                      type: string
                  created:
                    type: string
              example:
                id: "s_a1b2c3d4e5f6"
                scope: ""
                model: "claude-sonnet-4-20250514"
                skills: ["web-search"]
                created: "2026-03-24T10:00:00Z"
    get:
      summary: List Chat Sessions
      description: Returns chat sessions for the authenticated user.
      operationId: listChatSessions
      tags: [Chat]
      parameters:
        - name: scope
          in: query
          description: Filter by scope. Omit to list all scopes.
          schema:
            type: string
      responses:
        '200':
          description: Array of chat sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    scope:
                      type: string
                    model:
                      type: string
                    title:
                      type: string
                    updated_at:
                      type: string
  /api/v1/chat/sessions/{id}/messages:
    post:
      summary: Send Message (SSE Stream)
      description: |
        Send a message in a chat session. Returns a Server-Sent Events (SSE) stream.

        **SSE Events:**
        - `connected` — session connected
        - `chunk` — streaming text chunk from the AI
        - `tool_call` — AI is calling a tool
        - `error` — an error occurred (includes `code`, `error`, `hint` fields)
        - `done` — response complete (includes `result`, `model`, token usage)

        **Common error codes in SSE:**
        - `NO_AI_KEY` — no API key configured (hint tells you how to set one)
        - `NO_AI_KEY_FOR_MODEL` — no key for the requested model
      operationId: sendChatMessage
      tags: [Chat]
      parameters:
        - name: id
          in: path
          required: true
          description: Chat session ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
            example:
              message: "What is the weather today?"
      responses:
        '200':
          description: SSE stream of AI response
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: Message is required
        '404':
          description: Session not found
  /api/v1/models:
    get:
      summary: List Available Models
      description: |
        Returns all known LLM models with their provider, context window, and pricing.
        Use `?enabled=true` to filter to only your enabled models.
      operationId: listModels
      tags: [Settings]
      parameters:
        - name: enabled
          in: query
          description: If "true", only return models enabled for your account
          schema:
            type: string
      responses:
        '200':
          description: Array of models
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                    provider:
                      type: string
                    context_window:
                      type: integer
                    input_cost_per_1m:
                      type: number
                    output_cost_per_1m:
                      type: number
  /api/v1/skills:
    get:
      summary: List Skills
      description: Returns all skills installed by the authenticated user.
      operationId: listSkills
      tags: [Skills]
      parameters:
        - name: q
          in: query
          description: Search filter
          schema:
            type: string
      responses:
        '200':
          description: Array of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
  /api/v1/skills/install:
    post:
      summary: Install Skill
      description: |
        Install a skill from Git URL, registry reference, or raw SKILL.md content.

        **From Git URL:** `{"source": "git", "url": "github.com/user/repo"}`
        **From content:** `{"source": "local", "content": "---\nname: my-skill\n..."}`
        **Two-phase (preview):** `{"mode": "preview", "url": "github.com/user/repo"}`
        **Two-phase (confirm):** `{"mode": "confirm", "session_id": "xxx", "skill_names": ["skill1"]}`
      operationId: installSkill
      tags: [Skills]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source:
                  type: string
                  enum: [local, git]
                content:
                  type: string
                  description: SKILL.md content (source=local)
                url:
                  type: string
                  description: Git URL or owner/repo (source=git)
                mode:
                  type: string
                  enum: [preview, confirm]
                session_id:
                  type: string
                skill_names:
                  type: array
                  items:
                    type: string
            examples:
              fromGit:
                summary: Install from GitHub
                value:
                  source: "git"
                  url: "github.com/user/skills-repo"
              fromContent:
                summary: Install from SKILL.md content
                value:
                  source: "local"
                  content: "---\nname: my-skill\ndescription: My custom skill\n---\nYou are a helpful assistant."
      responses:
        '201':
          description: Skill installed
        '400':
          description: Invalid request
  /api/v1/skills/install-zip:
    post:
      summary: Install Skill from ZIP
      description: Upload a ZIP file containing one or more SKILL.md files.
      operationId: installSkillZip
      tags: [Skills]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: ZIP file (max 50MB)
      responses:
        '201':
          description: Skills installed
        '400':
          description: Invalid file
  /api/v1/skills/{id}:
    delete:
      summary: Delete Skill
      description: Remove a skill from your workspace.
      operationId: deleteSkill
      tags: [Skills]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Skill deleted
        '404':
          description: Skill not found
...
  /api/v1/user/settings/ai-key/{provider}:
    delete:
      summary: Delete AI Key
      description: Remove your API key for a specific provider.
      operationId: userDeleteAIKey
      tags: [Settings]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Key deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via POST /api/v1/auth/login
...