# internal/server/server.go
package server

import (
	"context"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"os"
	"os/exec"
	"path/filepath"
	"strconv"
	"strings"
	"sync"
	"syscall"
	"time"
	"tofi-core/internal/bridge"
	"tofi-core/internal/daemon"
	"tofi-core/internal/chat"
	"tofi-core/internal/crypto"
	"tofi-core/internal/executor"
	"tofi-core/internal/models"
	"tofi-core/internal/skills"
	"tofi-core/internal/storage"
	"tofi-core/internal/workspace"
)

type Config struct {
	Port    int
	HomeDir string
}

// HoldSignal is sent through a hold channel to resume a paused agent
type HoldSignal struct {
	Action string // "continue" or "abort"
	Answer string // user's answer text (for tofi_ask_user)
}

// PreviewSession 缓存已 clone 但尚未确认安装的 skill 预览
type PreviewSession struct {
	Skills    []*models.SkillFile
	Source    *skills.ParsedSource
	Cleanup   func()
	CreatedAt time.Time
	Scope     string // "public" (git) or "private" (zip upload)
	UserID    string // installer user ID (used when Scope="private")
}

type Server struct {
	config   Config
	db       *storage.DB
	executor executor.Executor // Sandbox command executor
	// Hold channel management for agent tofi_suggest_install blocking
	holdMu       sync.Mutex
	holdChannels map[string]chan HoldSignal // cardID → signal channel

	// SSE event hubs: one per in-flight chat agent run. Lets multiple HTTP
	// clients (e.g. the original POST /messages + a GET /stream after refresh)
	// subscribe to the same event stream with replay.
	sessionHubsMu  sync.Mutex
	sessionHubs    map[string]*sessionHub
	sessionCancels map[string]context.CancelFunc

	// Preview session cache for two-phase skill install
	previewMu       sync.Mutex
	previewSessions map[string]*PreviewSession // sessionID → session

	// App scheduler (min-heap + timer for scheduled app runs)
	appScheduler *AppScheduler

	// File-based workspace (source of truth for agents)
	workspace     *workspace.Workspace
	workspaceSync *workspace.Sync

	// Chat session store (XML files + SQLite index)
	chatStore *chat.Store

	// Bridge Manager (Telegram bidirectional chat)
	bridgeManager *bridge.ChatBridgeManager

	// Per-user rate limiter
	rateLimiter *RateLimiter

	// Access token for token-mode auth (read from config.yaml)
	accessToken string
}

func NewServer(config Config) (*Server, error) {
	// 初始化 JWT Auth
	InitAuth()

	// 初始化加密密钥。
	// 默认严格：必须由环境提供 TOFI_ENCRYPTION_KEY。只有 TOFI_DEV=1 时才
	// 允许 fallback 到公开的默认 key —— 否则生产部署漏设环境变量会让所有
	// 用户的 AI key 用全网已知的密钥加密（等同于明文存储）。
	encryptionKey := os.Getenv("TOFI_ENCRYPTION_KEY")
	if encryptionKey == "" {
		if os.Getenv("TOFI_DEV") == "" {
			log.Fatal("TOFI_ENCRYPTION_KEY is required. Set it in your environment, " +
				"or set TOFI_DEV=1 to use a known-default key for local development.")
		}
		encryptionKey = "tofi-default-encryption-key!!123" // 32 bytes, dev-only
		log.Println("⚠️  TOFI_DEV: using the default encryption key — never use in production")
	}
	if err := crypto.InitEncryption(encryptionKey); err != nil {
		return nil, fmt.Errorf("failed to initialize encryption: %v", err)
	}

	db, err := storage.InitDB(config.HomeDir)
	if err != nil {
		return nil, err
	}

	// 迁移明文 AI Key 到加密存储（启动时自动，一次性）
	if err := db.MigrateAIKeysToSecrets(); err != nil {
		log.Printf("⚠️  AI key migration warning: %v", err)
	}

	// Initialize sandbox executor (direct execution with software-level isolation)
	exec := executor.NewDirectExecutor(config.HomeDir)

	// Read rate limit RPM from settings (default 60)
	rpm := 60
	if rpmStr, err := db.GetSetting("rate_limit_rpm", ""); err == nil && rpmStr != "" {
		if v, err := strconv.Atoi(rpmStr); err == nil && v > 0 {
			rpm = v
		}
	}

	return &Server{
		config:   config,
		db:       db,
		executor: exec,
		holdChannels:    make(map[string]chan HoldSignal),
		sessionHubs:     make(map[string]*sessionHub),
		sessionCancels:  make(map[string]context.CancelFunc),
		previewSessions: make(map[string]*PreviewSession),
		chatStore:       chat.NewStore(config.HomeDir, db),
		rateLimiter:     NewRateLimiter(rpm),
	}, nil
}

// createSessionHub registers a new hub + cancel func for a session. Returns
// an error if a hub already exists — we disallow concurrent agent runs on
// the same session.
func (s *Server) createSessionHub(sessionID string, cancel context.CancelFunc) (*sessionHub, error) {
	s.sessionHubsMu.Lock()
	defer s.sessionHubsMu.Unlock()
	if _, exists := s.sessionHubs[sessionID]; exists {
		return nil, fmt.Errorf("session %s already has an active agent", sessionID)
	}
	hub := newSessionHub()
	s.sessionHubs[sessionID] = hub
	s.sessionCancels[sessionID] = cancel
	return hub, nil
}

// getSessionHub returns the in-flight hub for a session, or nil if no agent
// is currently running.
func (s *Server) getSessionHub(sessionID string) *sessionHub {
	s.sessionHubsMu.Lock()
	defer s.sessionHubsMu.Unlock()
	return s.sessionHubs[sessionID]
}

// removeSessionHub clears both hub and cancel registrations for a session.
// Call after the agent goroutine completes.
func (s *Server) removeSessionHub(sessionID string) {
	s.sessionHubsMu.Lock()
	defer s.sessionHubsMu.Unlock()
	delete(s.sessionHubs, sessionID)
	delete(s.sessionCancels, sessionID)
}

// cancelSession triggers the cancel func associated with an in-flight agent
// run. Used by POST /abort to stop the agent mid-stream even when the agent
// is not currently on a hold.
func (s *Server) cancelSession(sessionID string) bool {
	s.sessionHubsMu.Lock()
	defer s.sessionHubsMu.Unlock()
	cancel, ok := s.sessionCancels[sessionID]
	if !ok {
		return false
	}
	cancel()
	return true
}

// createHoldChannel creates a buffered channel for a card to wait on
func (s *Server) createHoldChannel(cardID string) chan HoldSignal {
	s.holdMu.Lock()
	defer s.holdMu.Unlock()
	ch := make(chan HoldSignal, 1)
	s.holdChannels[cardID] = ch
	return ch
}

// signalHold sends a signal to unblock the agent waiting on this card
func (s *Server) signalHold(cardID, action, answer string) bool {
	s.holdMu.Lock()
	defer s.holdMu.Unlock()
	ch, ok := s.holdChannels[cardID]
	if !ok {
		return false
	}
	ch <- HoldSignal{Action: action, Answer: answer}
	delete(s.holdChannels, cardID)
	return true
}

// removeHoldChannel cleans up a hold channel (used on timeout/cleanup)
func (s *Server) removeHoldChannel(cardID string) {
	s.holdMu.Lock()
	defer s.holdMu.Unlock()
	delete(s.holdChannels, cardID)
}

// --- Preview Session Management ---

func (s *Server) createPreviewSession(id string, session *PreviewSession) {
	s.previewMu.Lock()
	defer s.previewMu.Unlock()
	s.previewSessions[id] = session
}

func (s *Server) getPreviewSession(id string) *PreviewSession {
	s.previewMu.Lock()
	defer s.previewMu.Unlock()
	return s.previewSessions[id]
}

func (s *Server) removePreviewSession(id string) {
	s.previewMu.Lock()
	defer s.previewMu.Unlock()
	if sess, ok := s.previewSessions[id]; ok {
		if sess.Cleanup != nil {
			sess.Cleanup()
		}
		delete(s.previewSessions, id)
	}
}

func (s *Server) cleanupPreviewSessions() {
	ticker := time.NewTicker(2 * time.Minute)
	defer ticker.Stop()
	for range ticker.C {
		s.previewMu.Lock()
		now := time.Now()
		for id, sess := range s.previewSessions {
			if now.Sub(sess.CreatedAt) > 10*time.Minute {
				if sess.Cleanup != nil {
					sess.Cleanup()
				}
				delete(s.previewSessions, id)
				log.Printf("[skills] preview session %s expired, cleaned up", id)
			}
		}
		s.previewMu.Unlock()
	}
}

func (s *Server) Start() error {
	defer s.db.Close()

	// Load access token for token-mode auth
	s.loadAccessToken()

	// 初始化文件 workspace 并同步到 DB 索引
	s.initWorkspace()
	s.syncWorkspaceOnStartup()

	// 安装/更新 System Skills（内置技能）
	skills.InstallSystemSkills(s.db, s.config.HomeDir)

	// 统一恢复所有僵尸状态
	s.recoverAll()

	// 启动 preview session 清理 goroutine
	go s.cleanupPreviewSessions()

	// 启动 plan 文件 TTL 清理
	s.startPlanCleanup()

	// 启动 app run log/session TTL 清理
	s.startAppRunCleanup()

	// 启动 refresh token 清理
	s.startRefreshTokenCleanup()

	// 初始化 Stripe（可选 — 不配置则 checkout 端点返回 503）
	if initStripe() {
		log.Println("[stripe] initialized")
	}

	// 启动订阅过期检查（每小时）
	s.startSubscriptionExpiryChecker()

	// 启动 rate limiter 清理
	rateLimitStop := make(chan struct{})
	go s.rateLimiter.StartCleanup(rateLimitStop)
	defer close(rateLimitStop)

	// 启动 App 调度器（DB-poll based）
	s.appScheduler = NewAppScheduler(s)
	if err := s.appScheduler.Start(); err != nil {
		log.Printf("App Scheduler start failed: %v", err)
	}
	defer s.appScheduler.Stop()

	// Start Bridge Manager (Telegram bidirectional chat)
	dispatcher := bridge.NewDispatcher(s.db, s.chatStore, func(userID, scope string, session *chat.Session, message string, opts *bridge.ExecuteOptions) error {
		_, err := s.executeChatSession(userID, scope, session, message, nil, opts)
		return err
	})
	dispatcher.SetRestartFn(func(botToken, chatID string) {
		log.Printf("[Server] Restart requested via Telegram (chat %s)...", chatID)
		exe, exeErr := os.Executable()
		if exeErr != nil {
			log.Printf("[Server] Cannot find executable: %v", exeErr)
			return
		}
		// Write restart-notify file so the new process can send confirmation
		if botToken != "" && chatID != "" {
			notifyFile := filepath.Join(s.config.HomeDir, ".restart-notify")
			_ = os.WriteFile(notifyFile, []byte(botToken+"\n"+chatID), 0600)
		}
		// Fork detached shell: wait for us to die (release port), then start new daemon
		restartSh := fmt.Sprintf("sleep 2 && %s start --home %s --port %d", exe, s.config.HomeDir, s.config.Port)
		cmd := exec.Command("sh", "-c", restartSh)
		cmd.SysProcAttr = &syscall.SysProcAttr{Setsid: true}
		if startErr := cmd.Start(); startErr != nil {
			log.Printf("[Server] Failed to schedule restart: %v", startErr)
			return
		}
		log.Println("[Server] Restart scheduled, exiting now...")
		daemon.RemovePID(s.config.HomeDir)
		time.Sleep(500 * time.Millisecond)
		os.Exit(0)
	})
	s.bridgeManager = bridge.NewManager(s.db, dispatcher)
	s.bridgeManager.StartAll()
	log.Println("[Server] Bridge Manager started")
	defer s.bridgeManager.StopAll()

	// Check for pending restart notification
	s.sendRestartNotification()

	mux := http.NewServeMux()

	// CORS 中间件包装器
	corsHandler := func(next http.Handler) http.Handler {
		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
			// 设置 CORS 头
			w.Header().Set("Access-Control-Allow-Origin", "*")
			w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
			w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
			w.Header().Set("Access-Control-Max-Age", "3600")
			w.Header().Set("X-Frame-Options", "DENY")
			w.Header().Set("X-Content-Type-Options", "nosniff")
			w.Header().Set("X-XSS-Protection", "1; mode=block")
			w.Header().Set("Referrer-Policy", "strict-origin-when-cross-origin")

			// 处理预检请求
			if r.Method == "OPTIONS" {
				w.WriteHeader(http.StatusOK)
				return
			}

			next.ServeHTTP(w, r)
		})
	}

	// 公开路由
	mux.HandleFunc("GET /health", s.handleHealth)
	mux.HandleFunc("GET /api/v1/auth/setup_status", s.handleSetupStatus)
	mux.HandleFunc("POST /api/v1/auth/setup", s.handleSetupAdmin)
	mux.HandleFunc("POST /api/v1/auth/login", s.handleLogin)
	mux.HandleFunc("POST /api/v1/auth/register", s.handleRegister)
	mux.HandleFunc("POST /api/v1/auth/verify", s.handleVerifyEmail)
	mux.HandleFunc("POST /api/v1/auth/resend-code", s.handleResendCode)
	mux.HandleFunc("POST /api/v1/auth/refresh", s.handleRefreshToken)

	// 受保护的 API 路由 (包裹 AuthMiddleware)
	mux.HandleFunc("GET /api/v1/auth/me", s.AuthMiddleware(s.handleGetMe))

	// Global File Library
	mux.HandleFunc("GET /api/v1/files", s.AuthMiddleware(s.handleListFilesGlobal))
	mux.HandleFunc("POST /api/v1/files", s.AuthMiddleware(s.handleUploadFileGlobal))
	mux.HandleFunc("GET /api/v1/files/{id}/preview", s.AuthMiddleware(s.handlePreviewFileGlobal))
	mux.HandleFunc("DELETE /api/v1/files/{id}", s.AuthMiddleware(s.handleDeleteFileGlobal))

	// Secret 管理路由
	mux.HandleFunc("POST /api/v1/secrets", s.AuthMiddleware(s.handleCreateSecret))
	mux.HandleFunc("GET /api/v1/secrets", s.AuthMiddleware(s.handleListSecrets))
	mux.HandleFunc("GET /api/v1/secrets/{name}", s.AuthMiddleware(s.handleGetSecret))
	mux.HandleFunc("DELETE /api/v1/secrets/{name}", s.AuthMiddleware(s.handleDeleteSecret))

	// Skills 管理路由
	mux.HandleFunc("GET /api/v1/skills/system", s.AuthMiddleware(s.handleListSystemSkills))
	mux.HandleFunc("GET /api/v1/skills/collection", s.AuthMiddleware(s.handleGetCollection))
	mux.HandleFunc("DELETE /api/v1/skills/collection", s.AuthMiddleware(s.handleDeleteCollection))
	mux.HandleFunc("GET /api/v1/skills", s.AuthMiddleware(s.handleListSkills))
	mux.HandleFunc("GET /api/v1/skills/{id}", s.AuthMiddleware(s.handleGetSkill))
	mux.HandleFunc("POST /api/v1/skills/create", s.AuthMiddleware(s.handleCreateSkill))
	mux.HandleFunc("PUT /api/v1/skills/{id}", s.AuthMiddleware(s.handleUpdateSkill))
	mux.HandleFunc("POST /api/v1/skills/install", s.AuthMiddleware(s.handleInstallSkill))
	mux.HandleFunc("POST /api/v1/skills/install-zip", s.AuthMiddleware(s.handleInstallSkillZip))
	mux.HandleFunc("POST /api/v1/skills/{id}/export", s.AuthMiddleware(s.handleExportSkill))
	mux.HandleFunc("DELETE /api/v1/skills/{id}", s.AuthMiddleware(s.handleDeleteSkill))
	mux.HandleFunc("GET /api/v1/skills/{id}/resources", s.AuthMiddleware(s.handleGetSkillResources))
	mux.HandleFunc("PUT /api/v1/skills/{id}/resources", s.AuthMiddleware(s.handlePutSkillResource))
	mux.HandleFunc("DELETE /api/v1/skills/{id}/resources", s.AuthMiddleware(s.handleDeleteSkillResource))

	// Skills Registry (搜索 skills.sh 生态)
	mux.HandleFunc("GET /api/v1/registry/search", s.AuthMiddleware(s.handleRegistrySearch))

	// Chat Sessions

	mux.HandleFunc("POST /api/v1/chat/sessions", s.AuthMiddleware(s.handleCreateChatSession))
	mux.HandleFunc("GET /api/v1/chat/sessions", s.AuthMiddleware(s.handleListChatSessions))
	mux.HandleFunc("GET /api/v1/chat/sessions/{id}", s.AuthMiddleware(s.handleGetChatSession))
	mux.HandleFunc("DELETE /api/v1/chat/sessions/{id}", s.AuthMiddleware(s.handleDeleteChatSession))
	mux.HandleFunc("PATCH /api/v1/chat/sessions/{id}", s.AuthMiddleware(s.handleUpdateChatSession))
	mux.HandleFunc("GET /api/v1/chat/sessions/{id}/export", s.AuthMiddleware(s.handleExportSession))
	mux.HandleFunc("POST /api/v1/chat/sessions/{id}/messages", s.AuthMiddleware(s.handleChatMessage))
	mux.HandleFunc("GET /api/v1/chat/sessions/{id}/stream", s.AuthMiddleware(s.handleChatSessionStream))
	mux.HandleFunc("POST /api/v1/chat/sessions/{id}/continue", s.AuthMiddleware(s.handleChatSessionContinue))
	mux.HandleFunc("POST /api/v1/chat/sessions/{id}/abort", s.AuthMiddleware(s.handleChatSessionAbort))

	// Settings / AI Key 管理
	mux.HandleFunc("GET /api/v1/settings/ai-keys", s.AuthMiddleware(s.handleListAIKeys))
	mux.HandleFunc("POST /api/v1/settings/ai-keys", s.AuthMiddleware(s.handleSetAIKey))
	mux.HandleFunc("DELETE /api/v1/settings/ai-keys/{provider}", s.AuthMiddleware(s.handleDeleteAIKey))
	mux.HandleFunc("GET /api/v1/settings/preferred-model", s.AuthMiddleware(s.handleGetPreferredModel))
	mux.HandleFunc("POST /api/v1/settings/preferred-model", s.AuthMiddleware(s.handleSetPreferredModel))
	mux.HandleFunc("GET /api/v1/settings/enabled-models", s.AuthMiddleware(s.handleGetEnabledModels))
	mux.HandleFunc("POST /api/v1/settings/enabled-models", s.AuthMiddleware(s.handleSetEnabledModels))
	mux.HandleFunc("GET /api/v1/models", s.AuthMiddleware(s.handleListModels))

	// User-facing AI Key management
	mux.HandleFunc("PUT /api/v1/user/settings/ai-key", s.AuthMiddleware(s.handleUserSetAIKey))
	mux.HandleFunc("GET /api/v1/user/settings/ai-keys", s.AuthMiddleware(s.handleUserListAIKeys))
	mux.HandleFunc("DELETE /api/v1/user/settings/ai-key/{provider}", s.AuthMiddleware(s.handleUserDeleteAIKey))

	// Token revocation
	mux.HandleFunc("POST /api/v1/auth/revoke", s.AuthMiddleware(s.handleRevokeTokens))
	mux.HandleFunc("DELETE /api/v1/admin/users/{id}/tokens", s.AdminMiddleware(s.handleAdminRevokeUserTokens))

	// Spend cap management
	mux.HandleFunc("PUT /api/v1/user/settings/spend-cap", s.AuthMiddleware(s.handleSetUserSpendCap))
	mux.HandleFunc("GET /api/v1/user/settings/spend-cap", s.AuthMiddleware(s.handleGetUserSpendCap))
	mux.HandleFunc("PUT /api/v1/admin/settings/spend-cap", s.AdminMiddleware(s.handleSetSystemSpendCap))

	// API Key management
	mux.HandleFunc("POST /api/v1/user/api-keys", s.AuthMiddleware(s.handleCreateAPIKey))
	mux.HandleFunc("GET /api/v1/user/api-keys", s.AuthMiddleware(s.handleListAPIKeys))
	mux.HandleFunc("DELETE /api/v1/user/api-keys/{id}", s.AuthMiddleware(s.handleDeleteAPIKey))

	// Admin: allow_user_keys toggle
	mux.HandleFunc("PUT /api/v1/admin/settings/allow-user-keys", s.AdminMiddleware(s.handleSetAllowUserKeys))

	// Subscription / Plan
	mux.HandleFunc("GET /api/v1/user/subscription", s.AuthMiddleware(s.handleGetSubscription))
	mux.HandleFunc("GET /api/v1/user/subscription/usage", s.AuthMiddleware(s.handleGetUsage))
	mux.HandleFunc("POST /api/v1/user/subscription/checkout", s.AuthMiddleware(s.handleCreateCheckout))
	mux.HandleFunc("POST /api/v1/user/subscription/portal", s.AuthMiddleware(s.handleCreatePortal))
	mux.HandleFunc("GET /api/v1/user/subscription/events", s.AuthMiddleware(s.handleListSubEvents))
	mux.HandleFunc("POST /api/v1/webhooks/stripe", s.handleStripeWebhook)

	// Connectors — 统一多渠道 API
	mux.HandleFunc("GET /api/v1/connectors", s.AuthMiddleware(s.handleListConnectors))
	mux.HandleFunc("POST /api/v1/connectors", s.AuthMiddleware(s.handleCreateConnector))
	mux.HandleFunc("GET /api/v1/connectors/{id}", s.AuthMiddleware(s.handleGetConnector))
	mux.HandleFunc("DELETE /api/v1/connectors/{id}", s.AuthMiddleware(s.handleDeleteConnector))
	mux.HandleFunc("PUT /api/v1/connectors/{id}/toggle", s.AuthMiddleware(s.handleToggleConnector))
	mux.HandleFunc("POST /api/v1/connectors/{id}/verify", s.AuthMiddleware(s.handleConnectorVerify))
	mux.HandleFunc("GET /api/v1/connectors/{id}/verify-status", s.AuthMiddleware(s.handleConnectorVerifyStatus))
	mux.HandleFunc("GET /api/v1/connectors/{id}/receivers", s.AuthMiddleware(s.handleConnectorReceivers))
	mux.HandleFunc("DELETE /api/v1/connectors/{id}/receivers/{rid}", s.AuthMiddleware(s.handleDeleteConnectorReceiver))
	mux.HandleFunc("POST /api/v1/connectors/{id}/test", s.AuthMiddleware(s.handleConnectorTest))

	// App-Connector linking
	mux.HandleFunc("GET /api/v1/agents/{id}/connectors", s.AuthMiddleware(s.handleListAppConnectors))
	mux.HandleFunc("POST /api/v1/agents/{id}/connectors", s.AuthMiddleware(s.handleLinkAppConnector))
	mux.HandleFunc("DELETE /api/v1/agents/{id}/connectors/{cid}", s.AuthMiddleware(s.handleUnlinkAppConnector))

	// App 管理路由
	mux.HandleFunc("POST /api/v1/plans", s.AuthMiddleware(s.handlePlanApp))
	mux.HandleFunc("GET /api/v1/plans/{planId}", s.AuthMiddleware(s.handleGetPlan))
	mux.HandleFunc("POST /api/v1/plans/{planId}/approve", s.AuthMiddleware(s.handleApprovePlan))
	mux.HandleFunc("DELETE /api/v1/plans/{planId}", s.AuthMiddleware(s.handleDeletePlan))
	mux.HandleFunc("POST /api/v1/agents/parse-schedule", s.AuthMiddleware(s.handleParseSchedule))
	mux.HandleFunc("POST /api/v1/agents/manager/chat", s.AuthMiddleware(s.handleManagerChat))
	mux.HandleFunc("GET /api/v1/agents", s.AuthMiddleware(s.handleListApps))
	mux.HandleFunc("POST /api/v1/agents", s.AuthMiddleware(s.handleCreateApp))
	mux.HandleFunc("GET /api/v1/agents/{id}", s.AuthMiddleware(s.handleGetApp))
	mux.HandleFunc("PUT /api/v1/agents/{id}", s.AuthMiddleware(s.handleUpdateApp))
	mux.HandleFunc("DELETE /api/v1/agents/{id}", s.AuthMiddleware(s.handleDeleteApp))
	mux.HandleFunc("POST /api/v1/agents/{id}/activate", s.AuthMiddleware(s.handleActivateApp))
	mux.HandleFunc("POST /api/v1/agents/{id}/deactivate", s.AuthMiddleware(s.handleDeactivateApp))
	mux.HandleFunc("POST /api/v1/agents/{id}/run", s.AuthMiddleware(s.handleRunAppNow))
	mux.HandleFunc("GET /api/v1/agents/{id}/runs", s.AuthMiddleware(s.handleListAppRuns))
	mux.HandleFunc("GET /api/v1/agents/{id}/runs/{runId}", s.AuthMiddleware(s.handleGetAppRun))
	mux.HandleFunc("GET /api/v1/agents/{id}/runs/{runId}/session", s.AuthMiddleware(s.handleGetAppRunSession))
	mux.HandleFunc("GET /api/v1/agents/{id}/runs/{runId}/log", s.AuthMiddleware(s.handleGetAppRunLog))
	mux.HandleFunc("POST /api/v1/agents/{id}/runs/{runId}/abort", s.AuthMiddleware(s.handleAbortRun))
	mux.HandleFunc("POST /api/v1/agents/{id}/trigger", s.AuthMiddleware(s.handleTriggerApp))
	mux.HandleFunc("GET /api/v1/agents/{id}/stats", s.AuthMiddleware(s.handleGetAppStats))

	// Schedules
	mux.HandleFunc("GET /api/v1/schedules/upcoming", s.AuthMiddleware(s.handleGetUpcomingRuns))
	mux.HandleFunc("POST /api/v1/schedules/{runId}/skip", s.AuthMiddleware(s.handleSkipRun))

	// Admin 管理路由 (需要 admin 权限)
	mux.HandleFunc("GET /api/v1/admin/stats", s.AdminMiddleware(s.handleAdminGetStats))
	mux.HandleFunc("GET /api/v1/admin/settings/registration", s.AdminMiddleware(s.handleGetRegistrationSettings))
	mux.HandleFunc("PUT /api/v1/admin/settings/registration", s.AdminMiddleware(s.handleSetRegistrationSettings))
	mux.HandleFunc("GET /api/v1/admin/users", s.AdminMiddleware(s.handleAdminListUsers))
	mux.HandleFunc("POST /api/v1/admin/users", s.AdminMiddleware(s.handleAdminCreateUser))
	mux.HandleFunc("DELETE /api/v1/admin/users/{id}", s.AdminMiddleware(s.handleAdminDeleteUser))
	mux.HandleFunc("GET /api/v1/admin/secrets", s.AdminMiddleware(s.handleAdminListSecrets))
	mux.HandleFunc("DELETE /api/v1/admin/secrets/{id}", s.AdminMiddleware(s.handleAdminDeleteSecret))
	mux.HandleFunc("GET /api/v1/admin/usage", s.AdminMiddleware(s.handleAdminGetUsage))

	// API Documentation
	s.registerDocsRoutes(mux)

	// 配置 Server（应用 CORS 中间件）
	srv := &http.Server{
		Addr:         fmt.Sprintf(":%d", s.config.Port),
		Handler:      corsHandler(mux),
		ReadTimeout:  10 * time.Second,
		WriteTimeout: 30 * time.Second,
	}

	serverStartedAt = time.Now()
	log.Printf("🚀 Tofi Server listening on port %d", s.config.Port)
	return srv.ListenAndServe()
}

// Build info — set by main package at startup via SetBuildInfo().
var (
	buildVersion   = "dev"
	buildTime      = "unknown"
	buildCommit    = "unknown"
	serverStartedAt time.Time
)

// SetBuildInfo sets version metadata from ldflags (called once at startup).
func SetBuildInfo(version, commit, built string) {
	buildVersion = version
	buildCommit = commit
	buildTime = built
}

func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
	uptime := time.Since(serverStartedAt).Truncate(time.Second)

	totalApps, _ := s.db.CountAllApps()
	activeApps, _ := s.db.CountActiveApps()

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]any{
		"status":  "ok",
		"version": buildVersion,
		"commit":  buildCommit,
		"uptime":  uptime.String(),
		"apps": map[string]int{
			"total":  totalApps,
			"active": activeApps,
		},
	})
}

// sendRestartNotification checks for a .restart-notify file left by the previous
// process and sends a "restart complete" message to the Telegram user who requested it.
func (s *Server) sendRestartNotification() {
	notifyFile := filepath.Join(s.config.HomeDir, ".restart-notify")
	data, err := os.ReadFile(notifyFile)
	if err != nil {
		return // no pending notification
	}
	os.Remove(notifyFile)

	parts := strings.SplitN(string(data), "\n", 2)
	if len(parts) != 2 {
		return
	}
	botToken, chatID := strings.TrimSpace(parts[0]), strings.TrimSpace(parts[1])
	if botToken == "" || chatID == "" {
		return
	}

	sender := &bridge.TelegramSender{BotToken: botToken}
	_ = sender.SendMessage(chatID, "✅ Tofi 服务已重启完成")
	log.Printf("[Server] Sent restart confirmation to Telegram chat %s", chatID)
}